Concluindo pedidos

+++ src/main/java/br/com/fiap/lanchonete/domain/ports/repositories/PedidoRepositoryPort.java
@@ -13,7 +13,7 @@ public interface PedidoRepositoryPort {
 
 	Optional<Pedido> buscarPorCodigoPedido(String codigo);
 
-	Long ultimoPedido();
+	Integer ultimoPedido();
 
 	Pedido salvar(Pedido pedido);
 }
